requirements.txt
ping3
python-nmap
networkx
matplotlib